Fix FileDir weirdness
The generic `DirHandle` type points to an `IFS::FileDir` struct, which each filesystem defines internally as required.
It looks like the compiler sometimes mixes these up and causes problems,
though this has only been seen when compiling in debug mode for Host.

So instead, redefine `DirHandle` as an abstract type, move the `FileDir` definition inside the appropriate filesystem namespace
and add the `GET_FILEDIR()` macro to deal with checking and casting.